+++ src/main/java/com/codessquad/qna/controller/LoginController.java
@@ -1,0 +1,42 @@
+package com.codessquad.qna.controller;
+
+import com.codessquad.qna.domain.User;
+import com.codessquad.qna.service.UserService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+import javax.servlet.http.HttpSession;
+
+@Controller
+@RequestMapping("/users")
+public class LoginController {
+
+    private final UserService userService;
+    private final Logger logger = LoggerFactory.getLogger(UserController.class);
+
+    public LoginController(UserService userService) {
+        this.userService = userService;
+    }
+
+    @PostMapping("/login")
+    public String login(String userId, String password, HttpSession session) {
+        User findUser = userService.login(userId, password);
+        if (findUser == null) {
+            logger.debug("로그인 실패");
+            return "redirect:/users/loginForm";
+        }
+        logger.debug("로그인 성공");
+        session.setAttribute("sessionedUser", findUser);
+        return "redirect:/";
+    }
+
+    @GetMapping("/logout")
+    public String logout(HttpSession session) {
+        session.removeAttribute("sessionedUser");
+        return "redirect:/";
+    }
+}
