+++ src/main/java/com/codessquad/qna/controller/UserController.java
@@ -2,6 +2,7 @@ package com.codessquad.qna.controller;
 
 import com.codessquad.qna.domain.User;
 import com.codessquad.qna.service.UserService;
+import com.codessquad.qna.util.HttpSessionUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
@@ -15,7 +16,7 @@ import java.util.List;
 @RequestMapping("/users")
 public class UserController {
 
-    Logger logger = LoggerFactory.getLogger(UserController.class);
+    private final Logger logger = LoggerFactory.getLogger(UserController.class);
 
     private UserService userService;
 
@@ -45,52 +46,23 @@ public class UserController {
 
     @GetMapping("/{id}/form")
     public String renderUpdateForm(@PathVariable Long id, Model model, HttpSession session) {
-        User sessionedUser = (User) session.getAttribute("sessionedUser");
-        if (sessionedUser == null) {
+        if (!userService.checkSession(session, id)) {
             return "redirect:/users/loginForm";
         }
-        if (id != sessionedUser.getId()) {
-            throw new IllegalStateException("자신의 정보만 수정 가능");
-        }
-        User findUser = userService.findById(sessionedUser.getId());
-        model.addAttribute("user", findUser);
+        model.addAttribute("user", HttpSessionUtils.getUserFromSession(session));
         return "user/userUpdateForm";
     }
 
     @PutMapping("/update")
     public String userUpdate(User user, String newPassword, Model model, HttpSession session) {
-        User sessionedUser = (User) session.getAttribute("sessionedUser");
-        if (sessionedUser == null) {
+        if (!userService.checkSession(session, user.getId())) {
             return "redirect:/users/loginForm";
         }
-
-        if (sessionedUser.getId() != user.getId()) {
-            throw new IllegalStateException("자신의 정보만 수정 가능");
-        }
-
         if (userService.update(user, newPassword)) {
             return "redirect:/";
         }
         model.addAttribute("fail", true);
         return "user/userUpdateForm";
-    }
-
-    @PostMapping("/login")
-    public String login(String userId, String password, HttpSession session) {
-        User findUser = userService.login(userId, password);
-        if (findUser == null) {
-            logger.debug("로그인 실패");
-            return "redirect:/users/loginForm";
-        }
-        logger.debug("로그인 성공");
-        session.setAttribute("sessionedUser", findUser);
-        return "redirect:/";
-    }
-
-    @GetMapping("/logout")
-    public String logout(HttpSession session) {
-        session.removeAttribute("sessionedUser");
-        return "redirect:/";
     }
 
 }
